[heap] Clean up MainAllocator constructors
This CL unifies the MainAllocator constructors for main and background
threads by always taking a LocalHeap as the first argument.
allocation_info is now a pointer instead of a reference in order to
use it as an optional argument in the constructor.
Allocation observer and pending allocation support is dynamically
enabled for main thread LocalHeaps.

Having a common constructor for main and background threads also makes
it possible to unify HeapAllocator::SetupMain and
HeapAllocator::SetupBackground to remove this duplicated code.

While there exists a second constructor for MainAllocator, that one
is used by the GC only.

Bug: chromium:1480975
Change-Id: If69d3aa8ae1be97e5ffbd796dbeb0224c5489b11
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5062832
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91220}