[turboshaft] Introduce DisableValueNumbering scope
The BranchConditionDuplicationReducer intentionally adds certain
duplicate nodes to the graph, because that helps code generation.
Since the ValueNumberingReducer is running in the same phase, we
must prevent it from eliminating this duplication, to not defeat
its purpose. The new DisableValueNumbering scope accomplishes
that by temporarily disabling the ValueNumberingReducer.

Bug: v8:14108
Change-Id: Ifad5027dec37f3a3906308d85b343a10dda161cd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5062841
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91214}