[turboshaft] Refactor reducer stack to enable using it with other IRs
We would like to be able to reuse the reducer stack for others IRs (in
particular Maglev).

This CL extracts some Turboshaft-specific parts of the reducer stack
into a Turboshaft-specific bottom of the stack and top of the
stack. The middle of the stack should then be IR-independent. More CLs
will follow to make this happen.

Additionally, having a distinct top and bottom of the stack (Assembler
used to be both the top and the bottom) will be easier to replace just
the bottom of the stack, which could be useful to allow in-place
mutations.

Drive-by: rename OptimizationPhase to CopyPhase. This name is better
because OptimizationPhase is sometimes used for simple lowerings (ie,
not optimizations), and some optimizations (eg,
DecompressionOptimization) don't need an OptimizationPhase.

Bug: v8:12783
Change-Id: I75ae4a49259c0534cd9f2f7ddd6534bcf4da7821
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5054539
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91266}