[exceptions] Unify pending and scheduled exceptions
The difference wasn't very well understood, and actually leads to
wonky/broken semantics.

Scheduled exceptions were embedder-side exceptions, while pending
exceptions were used inside of v8. Conversions were done on the
boundaries. Pending exceptions could be cleared locally without
try/catch since we'd only propagate exceptions to external try/catch on
Promotion. An embedder-side TryCatch could be reused without
v8::TryCatch::Reset() first. If there already was an exception and a new
one was thrown, the exception was simply overwritten.

Or that was the theory.

V8 is reentrant. When the embedder returns to V8 a scheduled exception
had to be turned into a pending exception
(Isolate::PromoteScheduledException()). When V8 returns to the embedder,
a pending exception is turned into a scheduled exception
(Isolate::OptionalRescheduleException()) to make sure that exceptions
can be returned through multiple JS/embedder stack segments.

This CL drops the distinction between scheduled and pending exceptions.
A possible pending exception is dropped on rentry in V8, meaning the
embedder will always be able to call into V8 again despite being given
an error.

This slightly changes the semantics: if an embedder calls into V8 while
a pending exception was sitting around, it won't be "rethrown" on exit
to the caller of the embedder. However, this case (rentering V8 without
an explicit TryCatch) would already have been partially broken: if V8
would call out to the embedder again, on return to V8 the exception
would have been throw there, instead of only when returning to the first
call to V8.

Embedder code may need to be fixed to account for this change. A similar
change was needed in V8 in src/d8/async-hooks-wrapper.cc: Multiple calls
to SET_HOOK_FN needed to complete before a possible exception was
propagated to the caller. The solution is to wrap the calls in TryCatch
with ReThrow.

In a follow-up I'll look into not clearing Isolate::pending_exception on
entry at all, and instead always clear it when the exception is
propagated to TryCatch (as opposed to JS). This would change DCHECKs in
V8 that guarantee that we exit V8 when we have a pending exception. I'll
support those by adding debugmode only flags.

Bug: v8:7235
Change-Id: Iff7c8fb8ff2f0bddbac0d84ce832086fba330dd1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5050065
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91288}