[inspector] Add V8InspectorSession::Evaluate
This CL adds a stripped down version of Runtime#evaluate to the
inspector session. The main motiviation is as a replacement for
Page#addScriptToEvaluateOnNewDocument. It requires the "command line
API", which is a V8 inspector thing.

We want to change how the command line API injection works in the
future and we don't want to duplicate the work in blink, so
Page#addScriptToEvaluateOnNewDocument (and friends) should go
through the V8 inspector instead.

R=bmeurer@chromium.org

Bug: chromium:1069486, chromium:1505767, chromium:1509340
Change-Id: Ia50a455a145beb2c6077f44e2f8326699de480e6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5103511
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91429}