[maglev] Ensure ToString does not introduce unnecessary conversions
If the value is passed as is (since it is of type kString), then also
no conversion node is needed.

This only happens when the input is not tagged and still has type
string -- in other words in dead code.

Fixed: chromium:1508915

Change-Id: I86255a9292e734c3d16509ca4d58d179c046ff97
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5110560
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91453}