[sandbox] Use CodeWrappers for js-to-wasm wrappers
For moving Code objects out of the sandbox, they must be referenced
through code pointers from inside the sandbox. As the js-to-wasm
wrappers are currently kept in a WeakFixedArray, we cannot directly use
code pointers, so instead we reference them through the Code's
in-sandbox CodeWrapper object.

Bug: chromium:1500887
Change-Id: I0ecbc8de7c58a4e11d5cafcc2942d928a7c61f14
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5116680
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91486}