[turboshaft] Do not run LoopUnrolling when there are no loops to unroll
Running LoopUnrollingPhase when no loops can be unrolled is a bit of a
waste, as it basically does an identical copy of the graph.

With this CL, LoopUnrollingPhase first runs the loop analyzer, and
only if some loops can be unrolled, the copy (and unrolling) is
triggered.

Bug: v8:12783
Change-Id: Ia6391d64a664a887b70c8fd5ec08d2b694122f1a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5115540
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91483}