Restrict bit_cast
Using bit_cast on pointers does not accomplish anything useful, and
obscures what might be incorrect behavior. Do not allow base::bit_cast
to be misused in that way.

Port the checks present in https://crrev.com/c/5094596 to this repo's
version of bit_cast.

A previous version of this CL was landed in https://crrev.com/c/5116901
but had to be reverted because it didn’t compile with GCC 9. This
version of the CL has been fixed to work with that release.

Bug: chromium:1506769
Change-Id: I9177f47ef37263d7b361d1b8dd9bc6b1bb33d118
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5122432
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91526}