[turboshaft][wasm] Fix InstanceCache with non-returning inlinees
It's possible that an inlined function never returns to the caller,
typically because of an unconditional trap or exceptions. When this
happens, we should try to record Phi inputs for the InstanceCache from
this inlined function body.

Bug: v8:14108
Change-Id: I02434c4b58a8144bf7ead4033dcd3db7f62b01d6
Fixed: chromium:1511124
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5126073
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91550}