[maglev] Fix caching JSArray.length
We need to take into account that the length might be updated by
UpdateJSArrayLength if the index is greater than (or equal to) the
current JSArray length.
This CL turns UpdateJSArrayLength into a ValueNode to return the updated
length and records the ValueNode as the known length property.

Bug: v8:7700
Change-Id: Ic7ec26e3886bc34ac7c62e83ab39e0117bd40672
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5121905
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91572}