[codegen] Load trusted pointer table via root register
When compiling isolate-independent code, we do not have an isolate
available. Trying to construct an ExternalReference for the nullptr
isolate will thus crash.

We already have a `CHECK(root_array_available_)` there, so we can
instead directly load the trusted pointer table via the root register.

This CL also demotes the CHECK to a DCHECK because assembler options are
set per compiler, so we have very good coverage for this and don't need
a release-mode CHECK.

R=ahaas@chromium.org
CC=saelo@chromium.org

Bug: v8:14499
Change-Id: I650963a955f86ee0b9027d8726e0491fc1783510
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5132815
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91601}