[sandbox] Use soft abort mode for sandbox fuzzing
With this CL, the "sandbox fuzzing mode" (currently enabled by enabling
the sandbox crash filter, but in the future we'll probably want a
--sandbox-fuzzing flag) now enables the "soft" abort mode, which ignores
DCHECK failures and cleanly exits on CHECK/FATAL errors. In addition,
when the "soft" abort mode is enabled we now explicitly state that the
failure is "harmless". This both makes it clear that this is not a
security issue and causes fuzzers that pattern match on the stderr
output to ignore these cases.

Bug: chromium:1510306
Change-Id: I7fae3b5ef82ef326538320bb0cf43c404ddb6c55
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5135295
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91602}