[wasm] Move exception reachability analysis to the decoder
We currently have an reachability analysis which determines if a given block can throw, and if not, no code is generated for the respective
catch. However, this analysis is done in the separate decoder
interfaces. This resulted in the analyses sometimes not being
consistent. Specifically for Turboshaft, this analysis cannot be made
to be consistent with Turbofan and Liftoff. This is because Turboshaft
inlines during graph generation; a call is always considered
potentially throwing, but the inlined body might be not throwing.
We cannot afford this inconsistency because Turboshaft relies on
having the same feedback slots for function calls as Liftoff.
To solve this, we move the analysis from the individual interfaces to
the decoder.

Bug: chromium:1511849
Change-Id: If71e682edd6bb91d8c6efd31503715facb6fdc4b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5134849
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91672}