[sandbox] Introduce CodeEntrypointTag
When the sandbox is enabled, a Code's entrypoint pointer is stored in
the code pointer table (CPT). Then, any code entrypoint exposed in the
CPT can be invoked by an attacker from any place that does a control
flow transfer through the CPT. This can be problematic if the calling
convention differs between caller and callee. For example, bytecode
handlers are currently available through the CPT, but do not expect to
be called in the context of e.g. a JavaScript function, and will
misbehave if they are.

This CL now introduces a CodeEntrypointTag enum to solve this issue: the
tag is XORed into the top bits of the entrypoint pointer, and the
callsite must then untag the pointer with the expected tag. If the tags
mismatch, the pointer will be invalid.

For now, there are only two tags: the default tag and the bytecode
handler tag. As we never expect to call into a bytecode handler through
the CPT, no callsite needs to change as part of this CL. However, in the
future we'll likely want to distinguish between e.g. Wasm, RegExp, and
JavaScript code, at which point the respective callsites will need to
performing the untagging.

Bug: chromium:1507800
Change-Id: If0c4c01b402dab631a04213cb7ef588a7d201763
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5130973
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91660}