[sandbox] Implement basic ProtectedPointer support
This CL implements basic support for protected pointers: pointers to
HeapObjects that cannot directly be modified by an attacker. When
accessing such a pointer, it can be assumed that neither the pointer
itself, nor the pointed-to object have been manipulated by an attacker.
This is because ProtectedPointers can only be used for references
between TrustedObjects, so that both the pointer and the pointed-to
object live in trusted space, outside of the sandbox.

This CL:
- Introduces a new slot type: ProtectedPointerSlot.
- Adds a new visitor routine: VisitProtectedPointer.
- Implements helper routines for declaring, defining, and iterating
  over protected pointers fields.
- Changes the (de)serializer to support protected pointers.
- Makes the garbage collector aware of protected pointers. For that, a
  new write barrier and a TRUSTED_TO_TRUSTED remembered set are added.

This CL only implements C++ runtime support for protected pointers.
Support for these pointers in e.g. CSA will be added later.

Bug: chromium:1505752
Change-Id: I59da2e3179e3e7edec33ed86212e6db2128ee637
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5079470
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91663}