[sandbox] Reference Code from OSR feedback vector slots via its wrapper
For moving Code objects out of the sandbox, they must be referenced
through a code pointer (an index into the code pointer table) from
inside the sandbox. This CL does so for the reference from OSR
FeedbackSlots. As these reference are weak and essentially part of a
FixedArray, we use the Code's in-sandbox wrapper object instead of a
direct code pointer.

Bug: chromium:1500887
Change-Id: I6546766a1d947891e8702b29751aec446be9bd83
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5106128
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91689}