[interpreter] Fix derived constructor return value check
Derived constructors should return `this` when they would
have returned undefined. This is currently implemented as an AST rewrite
during parsing, where `return e` is rewritten to

return (tmp = e, tmp === undefined ? this : tmp)

This rewrite is incorrect, because the undefined check should be
evaluated after the derived constructor returns. Rewriting it in the
return expression position causes it to be evaluated before returning, which is incorrect in the case of try-finally.

This CL removes the AST rewrite and implements the check during bytecode
generation, by compiling `return` statements in derived constructor as
jumps to an epilogue that checks the return value before actually
returning.

Fixed: v8:12085
Change-Id: I084f3ba67486b75b06e119674ad4569e7649cf20
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5171908
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91696}