[reclient] Make reclient work without download_remoteexec_cfg
The Chromium build instructions for using reclient don't say to use
download_remoteexec_cfg, and while that works in Chromium, it doesn't
in V8. The difference is that Chromium has two flags that run
configure_reclient_cfg, one for the case where download_remoteexec_cfg
is true and one for the case where it is false. V8 only had the one
where it is true.

This patch renames the existing `configure_reclient_cfgs` hook to
`download_and_configure_reclient_cfgs`, to reflect the fact that it
does downloads the configuration, and adds a new
`configure_reclient_cfgs` hook that does not require
download_remoteexec_cfg.

Change-Id: Ief3649b7f2beb7729bd6823dccb5fed963211b0b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5172450
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#91719}