[wasm] Skip alignment check for single-byte accesses
The alignment check is redundant if the access size is 1. Turboshaft and
TurboFan optimize out the code later, but Liftoff emits this useless
sequence:

                  [ alignment check
0x2b9e02b5a8a6    a6  8bfb                 movl rdi,rbx
0x2b9e02b5a8a8    a8  83e700               andl rdi,0x0
0x2b9e02b5a8ab    ab  85ff                 testl rdi,rdi
0x2b9e02b5a8ad    ad  0f8574000000         jnz 0x2b9e02b5a927  <+0x127>
                  ]

This CL avoid emitting the alignment check for single-byte accesses
in all compilers. This will save memory and compile time also for the
optimizing compilers.

R=dlehmann@chromium.org

Change-Id: I8ca9719248113cf57fd3a34dc0c70180613d4872
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5177987
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91737}