[api] Add use counters for shipped Wasm features
This adds use counters for six Wasm proposals where we tracked usage but
never populated use counters with that data.
In order to avoid the same mistake in the future we also add static
asserts to check that any staged or shipped feature has a corresponding
use counter.

The chromium-side wiring will be done in a follow-up.

R=jkummerow@chromium.org, mlippautz@chromium.org

Change-Id: Id44911a80e61ec85b0cd700cd3eba4e229f310d5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5180775
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91766}