[turboshaft][x64] Make sure Load/Store offset=0 for atomic Loads/Stores
The instruction selector assumes that the offset of Loads and Stores
in 0. We thus now normalize atomic Load/Stores on x64 during
LoadStoreSimplificationReducer.

Drive-by: when the Index is constant but cannot be merged into the
offset because it would overflow 32 bits, we now instead merge the
offset into the index (which is a WordPtr, while the offset is a
Word32), to avoid emitting an addition in the instruction selector.

Bug: v8:12783
Change-Id: I80d33b83ac210ae54ad5a0ef17f120c7b6953027
Fixed: chromium:1513580
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5148172
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91778}