[turboshaft][wasm] Fix inlining budget for wasm-gc modules
In Turbofan we compare:
  initial_size + inlined_size + candidate_size < budget
In turboshaft we compare:
  inlined_size + candidate_size < budget
(with the same budgets)

One of the metrics baked into the `budget` is that we always allow
inlining of up to 10% of the initial graph size.
Given the wrong comparison, this means that for Turboshaft we allow
inlining of at least 110% of the initial graph size meaning that huge
functions can get incredibly large.

This CL fixes these budgets.
For some wasm-gc benchmarks this results in a ~5% performance regression
as these benchmarks profit from aggressive inlining while improving
compile time by about 20%.
(Note that we have decided to take similar regressions in Turbofan
for a better compromise of compile time and runtime in commit d420e55
and commit fc9ab92. These numbers could and should be re-evaluated
given that Turboshaft has much better compile time, yet the current
budgets only make sense when comparing the correct numbers.)

Note that linear wasm modules (i.e. no wasm-gc types) are not affected
by this change as inlining is not enabled for them (yet).

Bug: v8:14108, chromium:1517795
Change-Id: Ibb0493b654cfa88d0cfc0efb212123179abfd762
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5197673
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91842}